Create grid with createGrid()
This project CAN'T work without its core aspect - the drawing grid.

createGrid() now makes said grid.

Instead of using the total number of boxes (the result of the number of
rows multiplied onto itself) and adding them to .grid-container to make
the grid - which would make it so that you'd have to do math on the JS
side - createGrid() simply creates the specified number of rows and just
adds the same number of boxes to .grid-row.

Doing this allows the use of CSS to properly style the grid as needed.

Nice. Simple. Easy.