Add CSS styles to make grid work as intended
The game now has its coloring grid.

Using flexbox, .grid-container, .grid-row and .grid-box flexes to allow
for the grid to appear as it is required to - decided by the initial 16x16
createGrid() and then later on by the user's desired grid size.

Zero math required on the JS side. All CSS.