Add method to get player's input
Up until now, the grid only showed a 16x16 grid.

getPlayerInput() now allows the game to take the player's desired input
via prompt and display it inside of the .grid-container without “breaking”
out of the container's set height and width. Note - the method does not
work as intended - .set-grid-btn is not tied to getPlayerInput() yet.

As for why 100 is the set limit of the number of rows:

Anything higher number inputs can mess with the player's computer -
affecting both their experience and how the grid is rendered.