Create .reset-grid-btn method and functionality
Making the ability to reset the grid boxes nice and simple without
having to refresh the page makes for a better player experience.

Prior to this commit, the only way to reset the colored grid boxes was
to refresh the page.

Now, .reset-grid-btn clears the background color styles added by
boxColorChanger().

All with the simple click of a button.