Fix bug with getPlayerInput()
Previously, when getPlayerInput() would run and the input given by the
player was either incorrect or empty, it would only alert the
player that something was wrong and then clear out the grid.

To avoid this, when getPlayerInput() is called and the input is either
incorrect or empty, it'll just evoke getPlayerInput() until a correct
input is give by the player.